+++ pkg/hub/api/teams/teams.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 	"strconv"
 
-	authContext "github.com/kobsio/kobs/pkg/hub/middleware/userauth/context"
+	authContext "github.com/kobsio/kobs/pkg/hub/auth/context"
 	"github.com/kobsio/kobs/pkg/hub/store"
 	teamv1 "github.com/kobsio/kobs/pkg/kube/apis/team/v1"
 	"github.com/kobsio/kobs/pkg/log"