+++ pkg/hub/api/users/users.go
@@ -3,7 +3,7 @@ package users
 import (
 	"net/http"
 
-	authContext "github.com/kobsio/kobs/pkg/hub/middleware/userauth/context"
+	authContext "github.com/kobsio/kobs/pkg/hub/auth/context"
 	"github.com/kobsio/kobs/pkg/hub/store"
 	dashboardv1 "github.com/kobsio/kobs/pkg/kube/apis/dashboard/v1"
 	userv1 "github.com/kobsio/kobs/pkg/kube/apis/user/v1"