+++ pkg/log/log_test.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"testing"
 
-	authContext "github.com/kobsio/kobs/pkg/hub/middleware/userauth/context"
+	authContext "github.com/kobsio/kobs/pkg/hub/auth/context"
 
 	"github.com/go-chi/chi/v5/middleware"
 	"github.com/stretchr/testify/require"