+++ pkg/satellite/middleware/user/user.go
@@ -5,7 +5,7 @@ import (
 	"encoding/json"
 	"net/http"
 
-	authContext "github.com/kobsio/kobs/pkg/hub/middleware/userauth/context"
+	authContext "github.com/kobsio/kobs/pkg/hub/auth/context"
 
 	"github.com/kobsio/kobs/pkg/middleware/errresponse"
 )