+++ plugins/plugin-azure/cmd/kubernetesservices.go
@@ -3,7 +3,7 @@ package azure
 import (
 	"net/http"
 
-	authContext "github.com/kobsio/kobs/pkg/hub/middleware/userauth/context"
+	authContext "github.com/kobsio/kobs/pkg/hub/auth/context"
 	"github.com/kobsio/kobs/pkg/log"
 	"github.com/kobsio/kobs/pkg/middleware/errresponse"