+++ plugins/plugin-azure/cmd/monitor.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 	"strconv"
 
-	authContext "github.com/kobsio/kobs/pkg/hub/middleware/userauth/context"
+	authContext "github.com/kobsio/kobs/pkg/hub/auth/context"
 	"github.com/kobsio/kobs/pkg/log"
 	"github.com/kobsio/kobs/pkg/middleware/errresponse"