+++ plugins/plugin-azure/pkg/instance/instance_mock.go
@@ -3,7 +3,7 @@
 package instance
 
 import (
-	context "github.com/kobsio/kobs/pkg/hub/middleware/userauth/context"
+	context "github.com/kobsio/kobs/pkg/hub/auth/context"
 	containerinstances "github.com/kobsio/kobs/plugins/plugin-azure/pkg/instance/containerinstances"
 
 	costmanagement "github.com/kobsio/kobs/plugins/plugin-azure/pkg/instance/costmanagement"